LayoutTests/fast/block/positioning/relayout-nested-positioned-elements-crash.html

<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div style="position:absolute">
    <span style="position:relative">
        <div style="position:absolute">
            <input id="hideMe"/>
        </div>
    </span>
</div>
<script>
description("This tests that we don't cause an assertion failure on relayout of nested positioned elements. This test pass if we don't cause an assertion failure.");
window.jsTestIsAsync = true;

window.setTimeout(runTest, 0); // For some reason we need the setTimeout() for this test to work.
function runTest()
{
    document.getElementById("hideMe").style.display = "none";
    window.setTimeout(finalizeTest, 0);
}

function finalizeTest()
{
    shouldBeEqualToString('document.getElementById("hideMe").style.display', 'none');
    finishJSTest();
}
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
